- fixed bugs in which a pointer was cast to an int before going into a function.   On some 64-bit machines (maybe all), int is only 32 bits, so the pointer is   corrupted by being cast to int and then back to a pointer.  Instead, change   these arguments to pointers.  Then when an int needs to be passed, use a   macro REGNO2PTR(n) to cast the int to a pointer.  This should be safe for   64-bit systems.

+++ bochs/fpu/fpu_entry.c
@@ -444,22 +444,22 @@ do_another_FPU_instruction:
 		  break;
 		case 4:         /* fsub */
 		  clear_C1();
-		  FPU_sub(LOADED|loaded_tag, (int)&loaded_data, control_word);
+		  FPU_sub(LOADED|loaded_tag, &loaded_data, control_word);
 		  break;
 		case 5:         /* fsubr */
 		  clear_C1();
-		  FPU_sub(REV|LOADED|loaded_tag, (int)&loaded_data, control_word);
+		  FPU_sub(REV|LOADED|loaded_tag, &loaded_data, control_word);
 		  break;
 		case 6:         /* fdiv */
 		  clear_C1();
-		  FPU_div(LOADED|loaded_tag, (int)&loaded_data, control_word);
+		  FPU_div(LOADED|loaded_tag, &loaded_data, control_word);
 		  break;
 		case 7:         /* fdivr */
 		  clear_C1();
 		  if ( st0_tag == TAG_Zero )
 		    partial_status = status1;  /* Undo any denorm tag,
 						  zero-divide has priority. */
-		  FPU_div(REV|LOADED|loaded_tag, (int)&loaded_data, control_word);
+		  FPU_div(REV|LOADED|loaded_tag, &loaded_data, control_word);
 		  break;
 		}
 	    }
@@ -948,22 +948,22 @@ do_the_FPU_interrupt:
                   break;
                 case 4:         /* fsub */
                   clear_C1();
-                  FPU_sub(LOADED|loaded_tag, (int)&loaded_data, control_word);
+                  FPU_sub(LOADED|loaded_tag, &loaded_data, control_word);
                   break;
                 case 5:         /* fsubr */
                   clear_C1();
-                  FPU_sub(REV|LOADED|loaded_tag, (int)&loaded_data, control_word);
+                  FPU_sub(REV|LOADED|loaded_tag, &loaded_data, control_word);
                   break;
                 case 6:         /* fdiv */
                   clear_C1();
-                  FPU_div(LOADED|loaded_tag, (int)&loaded_data, control_word);
+                  FPU_div(LOADED|loaded_tag, &loaded_data, control_word);
                   break;
                 case 7:         /* fdivr */
                   clear_C1();
                   if ( st0_tag == TAG_Zero )
                     partial_status = status1;  /* Undo any denorm tag,
                                                   zero-divide has priority. */
-                  FPU_div(REV|LOADED|loaded_tag, (int)&loaded_data, control_word);
+                  FPU_div(REV|LOADED|loaded_tag, &loaded_data, control_word);
                   break;
                 }
             }
